nets, socket.c: Fix node shutdown correctness.
Since most node handling threads are created detached,
it is not correct to try to join on them later, so we
now do a better job of keeping track of this so that
we don't erroneously try to join nodes with detached
threads or where somebody else will join instead, as
with net_ssh.

Also fixes a memory leak in net_ssh caused by replacing
the file descriptor with -1 before passing it to
ssh_event_remove_fd.

Fixes #6