net_imap: Improve remote STATUS efficiency, fix memory leaks.
This significantly increases the performance of the STATUS
command when used with remote mailboxes. Previously, a
separate connection was set up and torn down for each
STATUS command issued to a remote server. We now reuse
connections for STATUS the same way that we do for all
other commands. To do this, rather than reverting to the
actually selected mailbox after the STATUS command finishes,
we let the override mailbox stay active until the next command
is issued, at which time we tear it down if it's not a STATUS
command. This at least allows for a single connection to be
used if a client issues the STATUS command for multiple mailboxes
on a remote server.

Other IMAP/SMTP fixes enclosed:

* If a proxied IMAP connection was established
to a different server while a connection to
a different server already existed, the old
connection was not properly cleaned up.
This is now fixed.

* Additionally, a few cases where scandir was used
with possible breaks are cleaned up to prevent
existing and other possible memory leaks.

* A few small fixes are included for net_smtp to
not keep retrying delivery on a permanent failure.
Additionally, some conditionally compiled code
can automatically correct broken mail queue envelopes,
though this will not likely remain permanently.