socket: Simplify socket spawning and registration.
Most network modules follow a predictable module of
creating TCP sockets, spawning a listener thread,
and registering a network service, all with a 1:1
mapping. Apart from duplicating code, this is inefficient
since every network module is spawning its own TCP
listener thread, as most of these threads are idle most
of the time.

This adds a TCP "multilistener" that can handle listening
to any TCP sockets registered with it. Internally, it also
handles network protocol registration and unregistration.
This also greatly simplifies the process of creating a TCP
listener, and allows a single thread to be used for all
these listeners.

A small benefit of this change is that any listeners registered
with the multilistener do not begin accepting connections until
after the BBS is fully started.