curl: Fix invalid free on HTTP failure.
Due to a known quirk in curl.c, a corrupted
pointer was being passed to free(), causing
a segfault on failed HTTP responses. We now
fix the pointer before doing this, as we do
in normal cases already.

A unit test has been added to capture this
failure; however, it is not built normally
since making external HTTP requests is slow.