net_imap: Add LITERAL+ support for APPEND.
This adds support for non-synchronizing literals,
denoted by the LITERAL+ capability, for the APPEND command.
This allows an RTT to be avoided when upload messages.

This also fixes a bug whereby auto-saved messages by net_smtp
were not automatically marked as Seen, since currently the
filtering engines cannot apply flags to messages.