mail: Add support for multiple domains.
Until now, only the BBS's hostname was
supported as the email domain. Support
is now added for secondary domains.
Email can now be sent to and from multiple
local domains handled by the BBS.

This is not fully equivalent to a generic
mail server that can handle multiple domains
independently, since authentication is tied
to BBS accounts. The limitation of this
implementation are discussed in the sample
config (mod_mail.conf).