mod_mail: Increase efficiency of reading/writing uid file.
This converts the .uidvalidity file in each maildir folder
from an ASCII file into a binary file, for efficiency of
read and write operations. This remains backwards compatible
with existing ASCII .uidvalidity files.

Additionally, handling of this file is streamlined, e.g.
we no longer check if the file is readable or writable
before trying to open it, since this is ultimately redundant.
Interacting with this file is, for now, an extremely common
and frequent operation and it should now be a little more
efficient.