tests: Move IMAP/POP3 compatibility tests to separate test module.
This moves the IMAP/POP3 compatibility testing logic in
both test_imap and test_pop3 into its own combined but
separate module. This removes dependencies of "the other
mail protocol" from each respective test module and
allows the IMAP and POP3 test modules to be primarily
concerned with themselves only.

Additionally, the compatibility tests involve a
TESTLOCK operation that takes a significant amount
of time. This test was previously run at the beginning
of the IMAP tests, which was quite annoying when debugging
test failures that occured later in the test module.
Now that this operation is in its own module, debugging
is unencumbered by these prolonged operations.