net_ws: Add WebSocket server module
This adds a WebSocket module that builds on top
of the HTTP module and server to provide WebSocket
capabilities to applications. This network module
uses the libwss library, which was developed for
this purpose. Together, the library and this module
provide a WebSocket server that applications within
the BBS can use in a relatively easy way.

Because any applications using WebSockets will
depend on net_ws, which in turn depends on mod_http,
it is now possible to have multilevel module dependencies,
which previously were not handled in module.c. As such,
a slightly hacky workaround has been added there to
ensure that preloading works properly during autoload
for such modules.