mod_webmail: Add webmail backend module.
Jamie Zawinski's Law of Software Envelopment stated that
"Every program attempts to expand until it can read mail.
Those programs which cannot so expand are replaced by ones which can."

The BBS already has a mail server, so I suppose that that cycle is
not complete until it is also packaged with accompanying webmail software.
So, here it is. This is not by any means complete or working perfectly,
but it does work well for basic IMAP operations that any webmail client
would be expected to support, and represents a good "starting point".

Truth be told, the webmail backend here itself does not have many strict
dependencies on the rest of the BBS, and in theory could be independent
software. However, it does require a WebSocket server, and a prototype
of doing this in PHP proved to be a poor idea in practice, mainly because
there are no good, robust IMAP client libraries for PHP. They are ALL
incomplete, and they all perform terribly because they simply don't provide
an interface that can make efficient or effective use of the IMAP protocol.
That being said, I explicitly did not want to write my own IMAP client
for this purpose, and have settled on using libetpan for mod_webmail.
This is more recent than Mark Crispin's UW IMAP c-client library,
and in practice it works relatively well: it supports key features that
were desired (such as IMAP IDLE support), and it performs very well
(though it is a C library, this is probably more due to the fact that
it has a good interface, which allows making effective and efficient
use of the IMAP protocol via the library).

This webmail backend must be used with the (an) accompanying frontend
(at this time, there is only one, which is still written in PHP,
because it's good enough for what it needs to do, at this time.

This frontend is maintained separately, because it doesn't have anything
to do with the BBS, and in theory the frontend could have multiple implementations,
e.g. in the future, we might have a frontend version in C that can run in the BBS itself.
For now, this is not a priority, so it continues to be maintained in PHP.
The only interface between the frontend and the backend are the JSON payloads
sent over the WebSocket connection, and as long as the API matches on both ends,
the details of each are not relevant to the other.

In the future, the backend could also be maintained separately and split from the BBS,
but for now, its dependence on net_ws, net_http, and mod_http made it easy
to include as a BBS module itself. In practice, the backend will also be optimized
over time for optimizations supported by the BBS's IMAP server, for the best possible
experience using the local mail server. As such, it is beneficial, though not required,
to include it (at least the backend) as part of this project.

The webmail software is not limited to the BBS mail servers and can be used
with any mail server. IMAP, but not POP3, is supported (no surprises there).
SMTP is currently handled by the frontend.