net_imap: Fix setting INTERNALDATE on APPEND.
The INTERNALDATE was not being set on APPENDs previously,
which meant that messages moved between IMAP accounts would
lose their original INTERNALDATE. There were also some parsing
issues with the date; this is fixed and the date is now set.

Additionally, there is a fix in mod_webmail to use timegm
instead of mktime so that the proper UTC epoch is handed
out to the client.