net_ws: Fix WebSocket ping timeout.
Browsers require WebSocket timeouts to be under 5 minutes
(if not even less). Since we were directly using an application's
requested timeout (in this case, an IMAP IDLE timeout of just
under 30 minutes), the connection would eventually time out
(according to the client) and close. To prevent this from happening,
poll timeouts are now limited to be under this browser threshold.
Additionally, the logic to send a ping to the client is moved
to the timeout branch instead of inside a (now unreachable) activity branch.