logger.c: Make logging more efficient.
Since logging is pretty central to the BBS, and in previous
profiling could comprise a non-neglible portion of CPU usage,
this improves __bbs_log to make logging slightly faster.

Now, a stack-allocated buffer is now preferred for constructing
the log message, rather than jumping straight to vasprintf or
asprintf. This means that for most log messages (which are
relatively small), no dynamic allocations are needed.

Also, in normal usage, we no longer bother grabbing the termlock
mutex to read log_stdout, because this really has no material
consequence. When run under helgrind, we may want this enabled
to avoid warnings about it.