tls.c, nets: Fix modern compiler errors.
Fixes errors raised by gcc 12 and API compatibility
with OpenSSL 3. In particular, gcc caught that the
calculation to get the last byte in a buffer was
incorrect (and in fact, downright spurious). To
do this properly, the original buffer size is
now passed into the SAFE_FAST_COND_APPEND macros.