tls.c: Fix usage/nonusage of HAVE_OPENSSL
The HAVE_OPENSSL usage exists currently in
the anticipation of wanting to compile the
BBS without OpenSSL support, although this
is not currently done or testing.

However, the usage in tls.c was quite sloppy;
this fixes its usage so that it is at least
semantically correct. SSL* is still typedefed
and used publicly in other places, so if OpenSSL
support were really missing, this would need
to be dummied out as well.