pthreads: Improve threading safety.
This addresses a class of deadlocks that have
been observed a few times now wherein threads
will start getting deadlocked somewhere that,
semantically, it should not be possible to
deadlock (typically in __bbs_log in logger.c).
My hunch has long been that this is due to the
use of pthread_cancel/pthread_kill causing threads
to exit while they are holding a mutex, which basically
spells disaster for the entire program. This demonstrates
why use of pthread_cancel is highly discouraged.

This does not fully eliminate usage of pthread_cancel,
but it addresses the above issue by:

* Shutdown listeners by calling shutdown() and
  close(), and exit the listener normally, rather
  than using bbs_pthread_cancel_kill. Formerly,
  cancel and kill were required because shutdown()
  was not called on the socket. If shutdown() is called,
  then cancel and kill are no longer necessary.
  This allows listener threads to exit much more cleanly.

* Removing redundant locking in logger.c. fprintf
  already serializes writes from multiple threads,
  there's no need to add a layer of locking on top
  of that. This was added initially during early
  testing with helgrind to make helgrind happy, but
  there's no material benefit of doing that.

* Explicitly disable thread cancellability in
  "critical sections" where a thread should not
  disappear. This is really a workaround for now, not
  a permanent solution, but this is a start.