net_ssh: Create login failure events.
Currently, the BBS core will automatically
create login failure events and broadcast
them for consumption, but this is not done
for SSH/SFTP connections because no node
exists at the time of authentication, so
the built-in event creator cannot be used.
As a result, spam SSH/SFTP traffic was not
being seen at all by mod_events, which is
responsible for blocking spam IPs.

We now manually create a login failure event
and broadcast it so that SSH/SFTP login failures
are handled the same way failures for all other
protocols are.