net_irc: Restore compatibility with broken clients.
Fixes a regression caused by ad7ca99d83d53d0324a91c5699556b4a5ec51e3e
that broke compatibility with IRC clients that only
send a LF for end of message, rather than CR LF as
the RFCs dictate. In particular, Ambassador appears
to do this. Prior to using readline, we were more tolerant
of line endings so this restores that flexibility.

Also includes Makefile updates for mod_webmail that should
have been included in 0573a57771825669f770a62bc51d99c993202921.
Not sure how mod_webmail is compiling and linking in the CI without
this, but this is (or should be) necessary for that to work.