net_ssh: Fix hang on server-initiated disconnect.
Fixes SSH sessions hanging when the BBS closes a node.
We now detect the other end of the PTY (connected to
the BBS node that has now gone away) has disappeared
via a POLLHUP and set a flag to exit, ensuring a
clean and timely cleanup.

Also fixes an issue where the IP address of the client
was not available in certain cases. We now store the IP
address at the beginning of the connection and reuse
that later as needed (a roundabout mechanism here needed
because we may not have a node).