net_irc: Fix ChanServ replies dumping to STDOUT.
Because ChanServ operates as a special IRC user
without a BBS node, it does not currently have
(or need) any mechanism to receive replies to
messages that it sends. However, because the wfd
member of the IRC user struct was not explicitly
set, the default value "0" was being used, which
equates to STDOUT. This meant that whenever messages
are sent to the ChanServ user object, that message
is also printed out to STDOUT, which, apart from
being redundant, was never intended or desired.
This is now fixed.