net_imap: Translate STATUS=SIZE if needed.
Commit 7a43d8fcf8e7ca6f0648e50b27dba064d9db6472
added support for RFC 8438 STATUS=SIZE,
but this introduced an incompatibility with
the remote IMAP proxy functionality whereby
servers that don't support RFC 8438 will
be confused if we pass through STATUS commands
that include SIZE (the client will include it
since we advertise support for it). Therefore,
if the remote server doesn't support this
extension (which most currently do not),
we need to intercept this and translate the
command, calculate the size another way,
and then translate back the response again.
Therefore, clients always transparently receive
a RFC8438 response if they asked for one, regardless
if the remote server supported it or not.