mod_webmail: Fix premature timeout.
Timing out after 7 seconds isn't sufficient
as FETCH 1:* (SIZE) can take a long time
on large folders.

Also, we no longer refresh the message listing
for SEEN and UNSEEN commands, since the frontend
can unambigously do this by itself (it doesn't
need any new information; this was just a lazy
way of updating the message listing, but it's
more efficient to do this on the frontend).