net_imap: Avoid unnecessary DB queries.
Every time we look up a folder name that
begins with the Other Users prefix, we
look up the beginning portion of the mailbox
name to see if it's a BBS username (in which
case, the username is translated to a user ID
before doing the translation to a disk path).
This meant that that if there are lots of
"Other Users" mailboxes, such as mapped in
.imapremote, then a DB query is done for every
single folder. Ultimately, this is not only
inefficient but also unnecessary, since if a
mapping exists, the user almost certainly wants
that mapping, not another user's mailbox (which
if not shared would just error out).

Consequently, we now keep track of the beginning
prefix (up to the first hierarchy delimiter) of
all remote mailboxes, so we can easily look them
up in a list and avoid this DB call if it's known
to be remote. This also has the advantage of avoiding
a user registering with the name of a mapping a user
has and "hijacking" that translation, since now the
remote mailbox mappings take priority.