valgrind: Fix various memory issues.
Fixes a few memory leaks and uninitialized memory usage.

Also updates the valgrind suppression file to be more
comprehensive, as it has not been very thorough of late,
leading to large dumps of useless findings. This is now
cleaned up so everything related to reachable memory
in shared libraries (dlopen/dlclose) should now be
suppressed.