logger.c: Allow max debug level for log file.
Allow a maximum debug level to be configured for the
log file, which works in conjunction with the existing
debug level setting. This is intended for sysops who
may want to be able to view debug messages on the console
but not have them persist to disk (since they can quickly
fill up a lot of space).