net_imap: Refactor mail and utility functions.
This refactors some utility functions from
net_imap to string.c and range.c.

A mail_ function is also moved to mod_mail,
where it was originally supposed to be. At the time,
this was not easily possible due to dependencies on
functions only in net_imap, but now that they are in
the core, this limitation no longer exists.

Apart from this refactoring being logic, it
also helps to reduce the size of net_imap by
moving code that isn't directly related to IMAP
out of the module.