net_imap: Refactor FETCH response generation.
The process_fetch function has grown into a
gigantic mess. Apart from growing rather
ad hoc as needed to implement certain functionality,
making it hard to extend, it is still incomplete
(not all BODY requests are currently handled).

This refactoring is therefore being done now
to reduce the chances of bugs or regressions
creeping in when further work is done to
round out FETCH response completeness.