mod_webmail: Use LIST-STATUS if available.
If the IMAP server supports LIST-STATUS, take advantage
of it. This enhances the already existing support for
STATUS=SIZE. mod_webmail will try to be as efficient as
possible based on the combination of LIST-STATUS and
STATUS=SIZE support available. This minimizes RTTs
and allows the folder pane to populate as fast as possible.