net_imap: Cache remote mailbox sizes if needed.
This adds caching for the STATUS responses returned
by remote IMAP servers that do not support RFC 8438
STATUS=SIZE, which requires us to issue FETCH 1:*
to calculate the size of each folder. This is very
slow and inefficient (it can take several seconds
or more on large mailboxes), so this operation should
be avoided unless necessary. We now only do this
if we don't have a cached size available to us or
the mailbox has changed since the cached size was
stored. This results in a significant speedup
with many/large mailboxes.