net_imap: Support UNAUTHENTICATE command.
Allows clients to logout and reuse a connection.
It would be nice if any major IMAP servers/providers
supported this extension for the client proxy to be
able to reuse connections, but as virtually none
do, supporting that when proxying to remote IMAP
servers would not add much value at this time,
unfortunately.