mod_webmail: Add deletion support.
Allow the \Deleted flag to be set on messages,
and allow an EXPUNGE operation to be issued.

This approach keeps in line with the IMAP
philosophy of deletion and the separation
between marking messages as deleted and
expunging a mailbox. The frontend may implement
its own user-friendlier abstraction on top of this.