mod_webmail: Emulate MOVE if needed.
Since MOVE is technically an IMAP extension,
not in the base IMAP4rev1 specification,
it is possible (though hopefully unlikely)
that there is some horrible IMAP server we
encounter that does not support MOVE.
In this case, fall back to doing a COPY instead
and then delete the original.

Technically, we should EXPUNGE as well, but
I'm a little uncomfortable with that, since
there could possibly be other messages in
the folder already with the Deleted flag set,
and the user might not want these expunged yet.
I think this is unlikely since typically the
deleted flag is only set on messages in a trash
mailbox, and you wouldn't typically move messages
from the trash somewhere else (unless it was
an accidental deletion).

In any case, I'll leave it to the user to manually
expunge for now.