net_ws: Internally handle ping poll interval.
Previously, it was possible for applications to
exceed the WebSocket ping threshold in certain
circumstances. Even though websocket_set_custom_poll_fd
limited the actual poll interval such that it fell
within the ping threshold, certain scenarios, such
as the application receiving activity on its file
descriptor multiple times and no application level
timeout occuring, would result in net_ws failing
to ping the client, since the timeout branch was
never being hit.

To fix this, we now ensure that the timeout branch
is always hit every ping interval, regardless of
the poll interval desired by the application. To
provide the illusion of polling for the application's
desired interval, we keep track of the elapsed time
and only trigger the timeout callback when the
application's threshold has been reached, rather
than whenever the timeout branch is hit.

This provides better separation of concerns since
applications don't need to be concerned
with the WebSocket ping interval at all,
and it's also more robust since, in theory,
we should no longer miss any pings.