tls.c: Remove misleading error message.
Removes an error message about not being able
to find an SSL structure in the linked list.
There is actually nothing wrong when this happens.
Analysis showed that the SSL structure can be
removed by ssl_unregister_fd just before this
call, in which case there is no need to mark it
as dead, because it's already gone. All the TLS
thread needs to do is cycle again and rebuild
its poll structures.