net_imap: Support LIST-STATUS for remote mailboxes.
For remote IMAP servers that support LIST-STATUS,
we now take advantage of that and store the STATUS
responses locally and just look things up there
when we need the STATUS of a remote mailbox (initially).
This allows us to cut down from N RTTs (where N = number
of mailboxes) to just 1 for getting STATUS information,
greatly improving performance when we're able to take
advantage of this.