net_imap: Send periodic IDLE notifications.
IMAP servers like Dovecot send periodic IDLE
notifications during IDLE as a sort of keepalive
mechanism. We now do so as well, and IDLE handling
is refactored to not require a state machine for
operation.

The default notify interval is 10 minutes,
which is greater than Dovecot's default as that
has widely been preceived to be too short.