net_imap: Prevent stack overflow while threading.
Fixes some stack overflow and use after free
bugs that could occur while threading.

This issue was exposed by the multiline parsing
that caused unrelated data to be treated as
"References" header content. The threader would
fail to find the dummy thread in question during
Step 3, and not be able to remove the child from
its parent. Additionally, attempts to dump threads
(with DEBUG_THREADING) could recurse indefinitely.

The following fixes are now in place:

* Explicit protections against stack overflow,
  and checks to see if recursion anomalies have occured.

* Fix bug in traversal logic in thread_remove.

* Fix header parsing that fills the References header var.