mod_webmail: Return attachment status in FETCHLIST.
The initial PHP webmail backend used a library that
returned attachment status, which was displayed in
the message list. This was lost in the conversion
to mod_webmail, but is restored now by using the
BODYSTRUCTURE FETCH item to determine if there
are any attachments in the message. This is preferred
to retrieving the entire message and doing the MIME
parsing, since attachment presence can be deduced
from BODYSTRUCTURE alone.