mod_mimeparse: Correct BODYSTRUCTURE body size.
The body size reported in the BODYSTRUCTURE
was previously just plain wrong, as it was
the size of the MIME headers, not the size of
the body itself.

This has been reworked using GMime APIs
to return the correct sizes. Because we
can't directly get the body size, we need
to calculate it ourself by taking the total
size and subtracting (carefully) the full
size of the headers.