mod_webmail: Fix sorted FETCHLIST ordering.
Ordering within a page of messages was
arbitrary in the response because although
the proper messages from the SORT response
were fetched, because the FETCH responses
can come from the IMAP server in any order,
we actually need to add them to the response
to the frontend in the same order of the request,
not merely going in the order of the response.

This is now done and order is properly maintained
regardless of the order of FETCH responses from the
server.