mod_http: Fix HTTPS redirects, flag bad requests.
Fix HTTPS redirects, which previously had 2 issues:
* If non-443 port was used, 443 was used instead of the right port.
* If an existing port was in the hostname, it was not removed.

Additionally, bad requests are now treated as auth failures.
In particular, this includes receiving a request where the
port in the hostname does not match the actual port, as well
as receiving CONNECT requests since we don't support those.
In both cases, such requests are almost guaranteed to be spam.