mod_webmail: Send empty FETCHLIST response for empty mailbox.
Previously, we were just bailing out of a FETCHLIST if
the mailbox had no messages, not sending any response.
This would cause the UI to not be updated in the frontend
and make it appear as though nothing had happened (e.g.
if a user had just deleted or moved all messages previously
in the folder). We now explicitly send an empty response
in order to update the frontend UI properly.