mod_webmail: Include unseen count in SELECT responses.
When a mailbox is selected, we now perform a STATUS
behind the scenes so we can include the unseen count
in the SELECT reply so it can update the number of
unread messages for that folder (and overall) in the
folder list.

The same mechanism is also used to update the folder
count after an EXPUNGE is completed.