mod_webmail: Fix header parsing, improve MIME decoding.
This fixes a bug with header parsing causing some headers
to be added more than once, with the result that the last
set would be the last line of the header value and override
the proper value.

Additionally, this improves the robustness of MIME decoding
for both headers and the message body. There is currently
an issue where, after decoding the header in mime_header_decode
or getting the body, the string contains unprintable characters
and thus json_string returns NULL since it jansson will refuse
to encode non UTF-8 compliant content.

As a somewhat hacky workaround that improves the robustness
of what we're able to handle, we remove any unprintable
characters in such cases and encode that to make it succeed.