net_imap: Explicitly read and discard ignored APPEND bodies.
If a client tries to append a message and it is
rejected, the entire message would still be
received if the client is using non-synchronizing
literals. However, we would just abort immediately
and thus the entire message was still in the buffer,
and we'd end up trying to process an entire RFC822
message as IMAP commands. This is not ideal for a
couple reasons:

* Large messages will spam the console and the logs
  A failed 5MB uploaded could easily result in
  5MB+ of logs in a second, if we're not careful.
* Even though most messages would fail since the
  message would not consist of valid IMAP commands,
  this is potentially a dangerous security issue as if a
  message contains lines that could be interpreted
  as IMAP commands, they would conceivably be
  executed. One can imagine a malicious message
  such as the following beind appended to a mailbox:

a1 STORE 1:* +FLAGS.SILENT (\Deleted)
a2 EXPUNGE

  This would have an extremely unpleasant side effect.

  To prevent such "append command injection", we now
  read and discard the message body if the client is
  using a non-synchronizing literal, to ensure it does
  not remain in the buffer when returning to process
  further commands, or disconnect the client otherwise.