net_imap: Return \Recent in FETCH, \Marked and \Unmarked in LIST.
We're already able to determine how many "Recent" messages a
mailbox has. We now have logic to determine if particular messages
should be considered Recent and return that as a message flag
in the FETCH response. If a mailbox has any recent messages, we
now also return the Marked flag; otherwise, we return Unmarked.

This will enable clients to be able to differentiate between
messages that are brand new and mailboxes containing such messages,
which can they then indicate in their user interface.