net_imap: Fix STATUS not being idempotent.
The STATUS command was not doing a read only traversal,
resulting in moving messages from the new to the cur
directory and thus perturbing the \Recent status of
messages. The traversal is now read only.

This does not seem to be a regression due to splitting
SELECT/EXAMINE and STATUS into separate functions; it
appears this bug has always existed.

As part of this fix, the traversal is split out into
a separate structure so that a traversal can occur
without permanently modifying the IMAP session data,
which makes it easier for STATUS to complete without
having any lingering side effects.