net_imap: Allow concurrent IMAP client proxies.
This is the first phase of a larger undertaking to
increase the efficiency of the IMAP client proxy
functionality in the IMAP server.

This patch moves all the client-related data to its
own structure that can be used independently of the
parent IMAP structure; most notably, there can now
be multiple client structures for a single IMAP session,
so there doesn't have to be an "at most one" relationship.

This can increase performance as when switching between remote
mailboxes on different remote IMAP servers, we no longer have
to shut down the first connection and open a second connection.
In fact, connections are generally no longer closed at all
during a user's session, unless the user hits the configurable
limit of concurrent client connections. This means that after
connections have been set up, we can switch between mailboxes
on different servers very quickly because the connections can
be reused, rather than repeatedly torn down and set up again
later.

Future optimizations are expected to revolve primarily
around taking advantage of multi-threading to perform
operations on multiple remote IMAP servers in parallel,
speeding up the time to generate a cumulative response
for the client.