net_imap: Eliminate temporary STATUS side effects.
Commit 3244bd508b0af3be76b7298cb673035e95f680a4
made most a STATUS operation idempotent by eliminating
most perturbation of the IMAP session structure when
a STATUS was performed. However, a few variables from
the IMAP structure were still used during the traversal.
This now completes this such that traversals don't
directly read or modify any IMAP session data. Any
copying between the traversal and IMAP session is done
before and/or after as necessary.

This allows for the STATUS operation to be implemented
much more cleanly now, because we don't have to worry
about restoring the previous data, since now we're not
perturbing that in the first place.