net_imap, tls.c: Improve TCP connection reuse.
This addresses a problem introduced by the recently
allowed parallel client proxy functionality that is
manifested in two forms: remote TLS connections that are
closed while they are not being serviced, and, at the
transport and application layers, a client connection
that has been closed or is no longer operational the
when we try to reuse it.

There are two major pieces to this:

* In tls.c, we no longer poll dead file descriptors.
  This is a big improvement because previously, we
  relied on the consumer of the TLS connection to
  service connections when they were closed in a
  relatively timely manner. In practice, doing this
  for IMAP client proxies would have meant polling
  all proxy connections as well as the actual IMAP
  client, significantly complicating handling. Even
  if this were done, it ignores the fact that connections
  are not always serviced immediately when closed, and
  the current design essentially leads to 100% CPU
  consumption until all dead connections are closed by
  the consumer, due to poll being level triggered and
  returning in perpetuity until the TLS connection is
  unregistered. Now, dead TLS connections are still including
  in the poll array, but their events bitmask is set to 0,
  so poll should not return activity on dead connections.
  This allows graceful toleration of delays in cleaning up
  dead TLS connections.

* Because we now hold on to remote server connections
  indefinitely, so that we can reuse them when possible,
  we need to be able to clean up a dead connection and
  establish a new connection if needed (transparently
  to the consumer). Checking if the file descriptor is
  still readable addresses this partly, but is not sufficient
  because some IMAP servers may only return failure after
  trying to use them. Thus, if enough time has passed that
  it's possible the connection no longer works, we explicitly
  send a NOOP command to verify that it's still functional,
  and if not clean up the connection and make a new one.

Another improvement here is that the buffer for client
readline operations is now included in the client structure
itself, so different functions don't need to have their
own stack allocated readline buffers and adjust pointers
accordingly. This simplifies the code a little bit, since
getting this wrong can lead to memory issues.

Finally, it's worth noting this patch took longer than
expected due to an issue with the IMAP connection returning
POLLHUP after reuse determined that a new connection needed
to be set up. The connection itself started fine (e.g.
CAPABILITIES response was successful) but soon would break
and return POLLHUP. This issue is fixed by simply moving
the order of an fclose call. This is still suspicious as
that doesn't provide a logical explanation currently, but
it consistently works correctly in the new order and previously
did not.