net_imap: Fix remote STATUS translation.
Fixes an issue with remote STATUS translation
potentially improperly removing the SIZE
item in translation, and make it suitable
for using in all remote STATUS commands, not
just those that involve the SIZE item.

While it was previously planned to fix the
passthrough issue of the remote mailbox name
not being translated back again, in reality,
it is probably desirable to always use the
remote_status function anyways, because it
internally handles caching of STATUS replies,
which can lead to transparent performance
improvements if we're talking to an efficient
server but the client is not efficient.