tls.c: Prevent constant polling of dead connections.
Commit 9e9469a9bd34f9b9664a7956c7a7a3285906bfce
attempts to fix the problem of TLS connections not
being cleaned up in a timely manner causing poll
to essentially busy wait in a tight loop, by setting
the poll events to 0. Unfortunately, this does not
actually work, because an events argument of 0
includes some implicit flags (see poll(2) man page).

In addition to that, we now set the file descriptor
to -1, which is an invalid fd that should not return
activity. This seems to work as intended, without
causing issues, and poll no longer monitors dead
connections after they are determined to be dead.